A workaround for the Chrome 10 where it cannot parse bad XML pages.
I really don't know if Facebook is the blame or a bug in Google.
somehow the parsing returns null and parse error wont allow the
DOM to parse correctly. JQuery has all the error handling built
into their magic $() so using that will parse irregular xml. But
it will throw an error every time, where the end result will not
be affected.